refactor: table component updated
