+++ docs/demos/table/DemoTableExtraColumn.vue
@@ -1,33 +1,7 @@
 <script setup lang="ts">
-const rows = [
-  {
-    name: 'Leanne Graham',
-    username: 'Bret',
-    website: 'hildegard.org',
-  },
-  {
-    name: 'Ervin Howell',
-    username: 'Antonette',
-    website: 'anastasia.net',
-  },
-  {
-    name: 'Clementine Bauch',
-    username: 'Samantha',
-    website: 'ramiro.info',
-  },
-  {
-    name: 'Patricia Lebsack',
-    username: 'Karianne',
-    website: 'kale.biz',
-  },
-  {
-    name: 'Chelsey Dietrich',
-    username: 'Kamren',
-    website: 'demarco.info',
-  },
-]
+import { rows } from './data'
 
-const columns = [
+const cols = [
   { name: 'name' },
   { name: 'username' },
   { name: 'actions' },
@@ -38,9 +12,9 @@ const columns = [
   <div class="cards-demo-container">
     <ATable
       :rows="rows"
-      :columns="columns"
+      :cols="cols"
     >
-      <template #row-actions>
+      <template #col-actions>
         <div class="flex items-center">
           <ABtn
             class="text-xs"
