+++ docs/demos/table/DemoTableSorting.vue
@@ -1,35 +1,19 @@
 <script setup lang="ts">
-const rows = [
-  {
-    name: 'Leanne Graham',
-    username: 'Bret',
-    website: 'hildegard.org',
-  },
-  {
-    name: 'Ervin Howell',
-    username: 'Antonette',
-    website: 'anastasia.net',
-  },
-  {
-    name: 'Clementine Bauch',
-    username: 'Samantha',
-    website: 'ramiro.info',
-  },
-  {
-    name: 'Patricia Lebsack',
-    username: 'Karianne',
-    website: 'kale.biz',
-  },
-  {
-    name: 'Chelsey Dietrich',
-    username: 'Kamren',
-    website: 'demarco.info',
-  },
+import { rows } from './data'
+
+const cols = [
+  { name: 'name' },
+  { name: 'username' },
+  { name: 'website', isSortable: false },
 ]
 </script>
 
 <template>
   <div class="cards-demo-container">
-    <ATable :rows="rows" />
+    <!-- Columns inferred from rows will have sorting enabled by default -->
+    <ADataTable
+      :cols="cols"
+      :rows="rows"
+    />
   </div>
 </template>
