+++ packages/anu-vue/src/components/data-table/events.ts
@@ -1,3 +1,3 @@
-import type { ServerTableEvents } from '@/components/server-table'
+import type { TableEvents } from '@/components/table'
 
-export interface DataTableEvents<HeaderClickCol> extends ServerTableEvents<HeaderClickCol> {}
+export interface DataTableEvents<HeaderClickCol> extends TableEvents<HeaderClickCol> {}
