+++ packages/anu-vue/src/components/index.ts
@@ -16,7 +16,6 @@ export { AMenu } from './menu'
 export { ARadio } from './radio'
 export { ARating } from './rating'
 export { ASelect } from './select'
-export { AServerTable } from './server-table'
 export { ASwitch } from './switch'
 export { ATable } from './table'
 export { ATextarea } from './textarea'
