--- packages/anu-vue/src/components/server-table/events.ts
@@ -1,3 +1,0 @@
-import type { TableEvents } from '@/components/table'
-
-export interface ServerTableEvents<HeaderClickCol> extends TableEvents<HeaderClickCol> {}
