--- packages/anu-vue/src/components/server-table/index.ts
@@ -1,4 +1,0 @@
-export { default as AServerTable } from './AServerTable.vue'
-export * from './events'
-export * from './props'
-
